feat: Enable in-browser JSX/TSX compilation
Integrates Babel Standalone script into index.html to allow for in-browser compilation of JSX and TSX. This simplifies development and testing by removing the need for a separate build step for these specific file types. The script tag for index.tsx has been updated to `text/babel` and `data-type="module"` to leverage this new functionality.